Adds current month to aboutForm
