+++ libs/financial-aid/shared/src/lib/formatters.ts
@@ -61,37 +61,6 @@ export const getActiveTypeForStatus: KeyMapping<ApplicationState, string> = {
   Approved: 'Approved',
 }
 
-// export type KeyMapping<TKey extends string, TValue> = { [K in TKey]: TValue }
-
-export const insertAt = (str: string, sub: string, pos: number) =>
-  `${str.slice(0, pos)}${sub}${str.slice(pos)}`
-
-export const formatPhoneNumber = (phoneNumber: string) => {
-  if (phoneNumber.length <= 10) {
-    return insertAt(phoneNumber.replace('-', ''), '-', 3) || '-'
-  }
-
-  return insertAt(phoneNumber.replace('-', ''), '-', 4) || '-'
-}
-
-export const formatNationalId = (nationalId: string) =>
-  insertAt(nationalId.replace('-', ''), '-', 6) || '-'
-
-export const months = [
-  'Janúar',
-  'Febrúar',
-  'Mars',
-  'Apríl',
-  'Maí',
-  'Júní',
-  'Júlí',
-  'Ágúst',
-  'September',
-  'Október',
-  'Nóvember',
-  'Desember',
-]
-
 export const aidCalculator = (
   homeCircumstances: HomeCircumstances,
   aid: {
