+++ libs/financial-aid/shared/src/lib/utils.ts
@@ -1,3 +1,5 @@
+import { months } from './const'
+
 export const getFileType = (fileName: string) => {
   return fileName.substring(fileName.lastIndexOf('.') + 1)
 }
@@ -5,3 +7,21 @@ export const getFileType = (fileName: string) => {
 export const getFileSizeInKilo = (file: { size?: number }) => {
   return Math.floor(file.size ? file.size / 1000 : 0)
 }
+
+export const currentMonth = () => {
+  return months[new Date().getMonth()].toLowerCase()
+}
+
+export const insertAt = (str: string, sub: string, pos: number) =>
+  `${str.slice(0, pos)}${sub}${str.slice(pos)}`
+
+export const formatPhoneNumber = (phoneNumber: string) => {
+  if (phoneNumber.length <= 10) {
+    return insertAt(phoneNumber.replace('-', ''), '-', 3) || '-'
+  }
+
+  return insertAt(phoneNumber.replace('-', ''), '-', 4) || '-'
+}
+
+export const formatNationalId = (nationalId: string) =>
+  insertAt(nationalId.replace('-', ''), '-', 6) || '-'
